pyproject.toml
[project]
name = "stegaboo"
version = "0.1.0"
description = "A playful image steganography CLI using Typer and Rich"
authors = [
    { name = "Justin Cox", email = "justincox.dev@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer[all]",
    "rich",
    "pillow"
]

[project.scripts]
stegaboo = "stegaboo.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"